+++ src/main/java/com/inttegro/orders/OrderPayoutFinancialAccount.java
@@ -2,6 +2,7 @@ package com.inttegro.orders;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.inttegro.bankaccounts.BankAccountConfig;
 import com.inttegro.customers.Address;
 import com.inttegro.customers.BillingDetails;
 import com.inttegro.customers.CustomerData;
@@ -17,7 +18,7 @@ import java.util.function.Consumer;
 public class OrderPayoutFinancialAccount {
     public FinancialAccountType type;
     public OrderPayoutWallet wallet;
-    @JsonProperty("bank_account") public com.inttegro.financialaccounts.BankAccountConfig bankAccount;
+    @JsonProperty("bank_account") public BankAccountConfig bankAccount;
     @JsonProperty("dosh_account") public Map<String, Object> doshAccount;
 
     public static Builder builder() { return new Builder(); }
@@ -26,7 +27,7 @@ public class OrderPayoutFinancialAccount {
         private final OrderPayoutFinancialAccount account = new OrderPayoutFinancialAccount();
         public Builder type(FinancialAccountType type) { account.type = type; return this; }
         public Builder wallet(OrderPayoutWallet wallet) { account.wallet = wallet; return this; }
-        public Builder bankAccount(com.inttegro.financialaccounts.BankAccountConfig bankAccount) { account.bankAccount = bankAccount; return this; }
+        public Builder bankAccount(BankAccountConfig bankAccount) { account.bankAccount = bankAccount; return this; }
         public Builder doshAccount(Map<String, Object> doshAccount) { account.doshAccount = doshAccount; return this; }
         public OrderPayoutFinancialAccount build() { return account; }
     }