+++ src/main/java/com/inttegro/paymentmethods/BankAccountSummary.java
@@ -1,7 +1,7 @@
 package com.inttegro.paymentmethods;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.inttegro.financialaccounts.BankAccountType;
+import com.inttegro.bankaccounts.BankAccountType;
 import java.util.Map;
 
 public class BankAccountSummary {