+++ src/main/java/com/inttegro/wallets/WalletConfig.java
@@ -1,7 +1,6 @@
-package com.inttegro.financialaccounts;
+package com.inttegro.wallets;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import java.util.Map;
 
 public class WalletConfig {
     public WalletType type;